+++ 2-foundation/common-error-fields/src/consts.ts
@@ -1,31 +1,44 @@
-export const STANDARD_ERROR_FIELDS = new Set<string>([
 
+export const STRICT_STANDARD_ERROR_FIELDS = new Set<string>([
 	// standard fields
 	// https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Error/prototype
 	'name',
 	'message',
+])
 
-	// quasi-standard
+export const QUASI_STANDARD_ERROR_FIELDS = new Set<string>([
+
+	// conv to array needed due to a babel bug 😢
+	...Array.from(STRICT_STANDARD_ERROR_FIELDS),
+
+	// quasi-standard: followed by all browsers + node
 	// https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Error/prototype
 	'stack',
 ])
 
 export const COMMON_ERROR_FIELDS = new Set<string>([
 
-	//...STANDARD_ERROR_FIELDS, // no, conv to array needed due to a babel bug 😢
-	...Array.from(STANDARD_ERROR_FIELDS.keys()),
+	// conv to array needed due to a babel bug 😢
+	...Array.from(QUASI_STANDARD_ERROR_FIELDS),
 
-	// standard in node
+	// standard in node only:
 	'code', // https://medium.com/the-node-js-collection/node-js-errors-changes-you-need-to-know-about-dc8c82417f65
 			  // https://nodejs.org/dist/latest/docs/api/errors.html#errors_node_js_error_codes
 
-	// non standard but widely used
+	// non standard but widely used:
 	'statusCode', // express, https://gist.github.com/zcaceres/2854ef613751563a3b506fabce4501fd
 	'shouldRedirect', // express, https://gist.github.com/zcaceres/2854ef613751563a3b506fabce4501fd
 	'framesToPop', // see facebook https://github.com/facebook/flux/blob/2.0.2/src/invariant.js
+])
 
-	// My (Offirmo) extensions
+export const COMMON_ERROR_FIELDS_EXTENDED = new Set<string>([
+
+	// conv to array needed due to a babel bug 😢
+	...Array.from(COMMON_ERROR_FIELDS),
+
+	// My (Offirmo) extensions:
 	'details',
 	'SEC',
 	'_temp', // used for passing state around between decorators
+	// TODO triage field? if needed.
 ])