+++ A-apps/online-adventur.es/functions/src/hello-world.ts
@@ -6,16 +6,17 @@ import {
 	NetlifyHandler,
 } from './sub/types'
 
+import { BUILD_DATE } from './sub/build'
 
 const handler: NetlifyHandler = async (
 	event: APIGatewayEvent,
 	badly_typed_context: Context,
 ): Promise<Response> => {
-	const msg = `Hello 20200718`
-	console.log('(log)', msg)
-	console.info('(info)', msg)
-	console.warn('(warn)', msg)
-	console.error('(error)', msg)
+	const msg = `Hello from ${BUILD_DATE}!`
+	console.log('(console.log)', msg)
+	console.info('(console.info)', msg)
+	console.warn('(console.warn)', msg)
+	console.error('(console.error)', msg)
 
 	return {
 		statusCode: 200,