+++ A-apps/online-adventur.es/functions/src/sub/services/sec.ts
@@ -82,4 +82,4 @@ if (ENV !== process.env.NODE_ENV) {
 
 /////////////////////////////////////////////////
 
-export { XError } from '../utils'
+export { LXXError } from '../utils'