+++ 2-foundation/practical-logger-node/CHANGELOG.md
@@ -4,6 +4,7 @@
 ## pending
 * TODO fix the "maximum call stack exceeded" when printing a SEC
 * TODO pretty+intelligently print the details
+* TODO publish private dependencies
 * [doc] ++
 * [chore] re-organized the source (monorepo) and tweaked the build script. No logic change.
 * [chore] bumped dependencies (minor)